Blocking Design

Assign arrays to samples with blocking experimental design by array slide. For example Agilent comes with an eight-plex design. That is the number of arrays in an array slide is eight; within each array slide, equal number of arrays are assigned to either sample group.

Usage

blocking.design(seed, num.per.slide = 8, num.array)

Arguments

seed

an integer used to initialize a pseudorandom number generator.

num.per.slide

number of arrays per array slide. It must be a multiple of 2. By default, num.per.slide = 8 for Agilent microarrays.

num.array

number of arrays. It must be a multiple of num.per.slide.

Value

a vector of array IDs in the order of assigning to samples that are assumed to be sorted by sample group of interest As a result, the first half of the array IDs are assigned to sample group 1 and the second half of the array IDs are assigned to sample group 2.

Examples

blocking.design(seed = 1, num.per.slide = 8, num.array = 128)
#>   [1]  15  80  76  17  39  71  29 126  53  92 110  46  94  38 128  60 114  48
#>  [19]  97  49  81  67  41  57  79 113  19 111  37  50  63  64  84  11  95 118
#>  [37]   4  24   1 123  30  55  31  82  21  88  10  66   8  78  89 107 103 124
#>  [55] 106   2 102  13  98  33  26  70  45 117  91  69 101  87  99  83 100 112
#>  [73] 108 120 116  93  54  14 105  40  47  32  36  72   5   3  75 119  25  52
#>  [91]  68 121 127  34  74  22  61   6  62  27  90  77  65 115  16  18  35  20
#> [109]  28   9  56 122 109  43 104  23  58  12   7  73  86  59  96  85  42 125
#> [127]  51  44