Plot misclassification error rates from PRECISION (both non-FLEX and FLEX) output

Plot error rates from PRECISION output that is being extracted by extract.precision.error

Usage

# S3 method for precision
plot(
  data,
  mytitle = "PRECISION results (misclass. error rates)",
  design.order = NULL,
  ylim = c(0, 0.5),
  iflex = TRUE
)

Arguments

data

a resulted PRECISION output from extract.precision.error.

mytitle

the title of the plot.

design.order

indices in the order of what user would like to plot from the resulted output from extract.precision.error. By default, every simulation is plotted.

ylim

y limit. By default, ylim = c(0, 0.5).

iflex

whether data if a PRECISION output from precision.simulate or precision.simulate.flex. By default, iflex = TRUE indicating that the data are from precision.simulate.flex.

Value

a boxplot of misclassification error rates

Examples

if (FALSE) {
## PRECISION output
plot.precision(data = precision.results.err.df,
  design.order = 1:10,
  mytitle = "My PRECISION results", iflex = FALSE)

## PRECISION FLEX output
plot.precision(data = precision.results.flex.err.df,
  design.order = 1:10,
  mytitle = "My PRECISION FLEX results")

# exclude ext.uh
plot.precision(data = precision.results.flex.err.df[!names(precision.results.flex.err.df) %in% "ext.uh"],
  design.order = 1:10,
  mytitle = "My PRECISION FLEX results")
}